fix(editor): move image geometry out of the node view file
react-refresh/only-export-components: ImageNode.tsx may only export the
component. ALIGN_STYLE and resizeWidth move to imageLayout.ts, which
extensions.ts and the tests now import directly — they never needed React.
